4-cinecla-daniel-manfredi-francesco/pyproject.toml
[project]
name = "tiny-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "focoos[onnx-cpu]",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-socketio>=5.3.0",
    "python-socketio>=5.11.0",
    "eventlet>=0.35.0",
    "requests>=2.31.0",
    "yt-dlp>=2024.0.0",
    "pillow>=10.0.0",
]

[tool.uv.sources]
focoos = { git = "https://github.com/FocoosAI/focoos.git" }
